+++ src/utils/toasting.ts
@@ -1,6 +1,5 @@
 import toast from "react-hot-toast";
 import { network } from "./metamask";
-import { OptionState } from "../models/option";
 
 const TOAST_DELAY = 333;
 
@@ -9,30 +8,6 @@ let showLoadingToast = false;
 const activeToasts : string[] = [];
 
 export const TOAST_DURATION = 2000;
-
-export const actionLabels =
-[
-    "published request"
-,   "withdrawn request"
-,   "opened option"
-,   "canceled option"
-,   "exercised option"
-];
-
-export const flavorLabels =
-[
-    "European"
-,   "American"
-];
-
-export const events =
-{
-    Published : OptionState.PUBLISHED
-,   Withdrawn : OptionState.WITHDRAWN
-,   Opened    : OptionState.OPEN
-,   Canceled  : OptionState.CANCELED
-,   Exercised : OptionState.EXERCISED
-};
 
 export function getIntervalString(value : string)
 {
